Fix error when >1 nfactors are requested (#42)
find.unsplit.Grps throws an error if more than 1 nfactor is requested. Specifically, the error is due to the following line.

inx <- which(unlist(lapply(ss, function(x, y) {all(y %in% x && x %in% y)

This was replaced with the determination of whether the items are in the list as two separate, and the unlisting as two separate commands.